Fixed noscript and slide in menu bugs
Noscript definitions added to ensure growing<hr />'s are visible even if
scripts are turned off. Also a noscript definition was added to make
only one of the swappable images in the image processing section,
visible.

Moreover a bug was fixed where the slide in menu would not close when
clicked, due to smooth scroll javascript.